requirements.txt
typer[all]==0.12.3
click==8.1.7
rich==13.7.1
requests==2.31.0
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
